refactor(server): decompose routes.ts monolith; honest 501s for lobotomized codegen endpoints
server/routes.ts drops from 1369 to 273 lines and becomes the umbrella
registrar. Extracted, preserving every public URL:

- server/routes/blueprints.ts (/api/blueprints/*) — ISA-88 CRUD + summary
- server/routes/vendors.ts (/api/{vendors,templates,data-types,controllers})
- server/routes/codegen.ts (/api/{generated-code,generate,ladder-logic})
- server/routes/assets.ts — real handlers replace the placeholder router
  that shadowed GET /api/assets (mounted before the real handler, it
  returned an empty list unconditionally; live bug, now fixed)

Truth-telling change: every /generate/* and /ladder-logic/* endpoint plus
blueprint import/seed had been stubbed with (null as any) after
server/blueprints/ was deleted in 6f9d9219c — they have thrown at runtime
ever since, returning misleading 500s. They now answer 501 Not
Implemented with a message naming the lost modules and #479 (restoration
issue). /api/ws/metrics and /api/ws/clients similarly stop calling the
deleted event-stream server and report only the live streams.

13 mount-path tests pin the public URLs, the 501 contract, and the
assets shadow-bug regression.

Fixes #446
